+++ src/ui/bind_handlers.js
@@ -33,6 +33,7 @@ module.exports = function bindHandlers(map: Map, options: {}) {
     el.addEventListener('mousedown', onMouseDown, false);
     el.addEventListener('mouseup', onMouseUp, false);
     el.addEventListener('mousemove', onMouseMove, false);
+    el.addEventListener('mouseover', onMouseOver, false);
     el.addEventListener('touchstart', onTouchStart, false);
     el.addEventListener('touchend', onTouchEnd, false);
     el.addEventListener('touchmove', onTouchMove, false);
@@ -57,7 +58,7 @@ module.exports = function bindHandlers(map: Map, options: {}) {
     }
 
     function onMouseUp(e: MouseEvent) {
-        const rotating = map.dragRotate && map.dragRotate.isActive();
+        const rotating = map.dragRotate.isActive();
 
         if (contextMenuEvent && !rotating) {
             // This will be the case for Mac
@@ -70,14 +71,23 @@ module.exports = function bindHandlers(map: Map, options: {}) {
     }
 
     function onMouseMove(e: MouseEvent) {
-        if (map.dragPan && map.dragPan.isActive()) return;
-        if (map.dragRotate && map.dragRotate.isActive()) return;
+        if (map.dragPan.isActive()) return;
+        if (map.dragRotate.isActive()) return;
 
         let target: any = e.toElement || e.target;
         while (target && target !== el) target = target.parentNode;
         if (target !== el) return;
 
         fireMouseEvent('mousemove', e);
+    }
+
+    function onMouseOver(e: MouseEvent) {
+
+        let target: any = e.toElement || e.target;
+        while (target && target !== el) target = target.parentNode;
+        if (target !== el) return;
+
+        fireMouseEvent('mouseover', e);
     }
 
     function onTouchStart(e: TouchEvent) {
@@ -126,7 +136,7 @@ module.exports = function bindHandlers(map: Map, options: {}) {
     }
 
     function onContextMenu(e: MouseEvent) {
-        const rotating = map.dragRotate && map.dragRotate.isActive();
+        const rotating = map.dragRotate.isActive();
         if (!mouseDown && !rotating) {
             // Windows: contextmenu fired on mouseup, so fire event now
             fireMouseEvent('contextmenu', e);