+++ src/ui/handler/drag_rotate.js
@@ -7,6 +7,7 @@ const browser = require('../../util/browser');
 
 import type Map from '../map';
 import type Point from '@mapbox/point-geometry';
+import type Transform from '../../geo/transform';
 
 const inertiaLinearity = 0.25,
     inertiaEasing = util.bezier(0, 0, inertiaLinearity, 1),
@@ -16,12 +17,6 @@ const inertiaLinearity = 0.25,
 /**
  * The `DragRotateHandler` allows the user to rotate the map by clicking and
  * dragging the cursor while holding the right mouse button or `ctrl` key.
- *
- * @param {Map} map The Mapbox GL JS map to add the handler to.
- * @param {Object} [options]
- * @param {number} [options.bearingSnap] The threshold, measured in degrees, that determines when the map's
- *   bearing will snap to north.
- * @param {bool} [options.pitchWithRotate=true] Control the map pitch in addition to the bearing
  */
 class DragRotateHandler {
     _map: Map;
@@ -29,13 +24,24 @@ class DragRotateHandler {
     _enabled: boolean;
     _active: boolean;
     _button: 'right' | 'left';
+    _eventButton: number;
     _bearingSnap: number;
     _pitchWithRotate: boolean;
+
+    _lastMoveEvent: MouseEvent;
     _pos: Point;
-    _startPos: Point;
+    _previousPos: Point;
     _inertia: Array<[number, number]>;
     _center: Point;
 
+    /**
+     * @param {Map} map The Mapbox GL JS map to add the handler to.
+     * @param {Object} [options]
+     * @param {number} [options.bearingSnap] The threshold, measured in degrees, that determines when the map's
+     *   bearing will snap to north.
+     * @param {bool} [options.pitchWithRotate=true] Control the map pitch in addition to the bearing
+     * @private
+     */
     constructor(map: Map, options: {
         button?: 'right' | 'left',
         element?: HTMLElement,
@@ -51,7 +57,8 @@ class DragRotateHandler {
         util.bindAll([
             '_onDown',
             '_onMove',
-            '_onUp'
+            '_onUp',
+            '_onDragFrame'
         ], this);
     }
 
@@ -98,23 +105,23 @@ class DragRotateHandler {
     }
 
     _onDown(e: MouseEvent) {
-        if (this._map.boxZoom && this._map.boxZoom.isActive()) return;
-        if (this._map.dragPan && this._map.dragPan.isActive()) return;
+        if (this._map.boxZoom.isActive()) return;
+        if (this._map.dragPan.isActive()) return;
         if (this.isActive()) return;
 
         if (this._button === 'right') {
-            const button = (e.ctrlKey ? 0 : 2);   // ? ctrl+left button : right button
-            let eventButton = e.button;
+            this._eventButton = e.button;
             if (typeof window.InstallTrigger !== 'undefined' && e.button === 2 && e.ctrlKey &&
                 window.navigator.platform.toUpperCase().indexOf('MAC') >= 0) {
                 // Fix for https://github.com/mapbox/mapbox-gl-js/issues/3131:
                 // Firefox (detected by InstallTrigger) on Mac determines e.button = 2 when
                 // using Control + left click
-                eventButton = 0;
+                this._eventButton = 0;
             }
-            if (eventButton !== button) return;
+            if (this._eventButton !== (e.ctrlKey ? 0 : 2)) return;
         } else {
             if (e.ctrlKey || e.button !== 0) return;
+            this._eventButton = 0;
         }
 
         DOM.disableDrag();
@@ -126,16 +133,18 @@ class DragRotateHandler {
 
         this._active = false;
         this._inertia = [[browser.now(), this._map.getBearing()]];
-        this._startPos = this._pos = DOM.mousePos(this._el, e);
+        this._previousPos = DOM.mousePos(this._el, e);
         this._center = this._map.transform.centerPoint;  // Center of rotation
 
         e.preventDefault();
     }
 
     _onMove(e: MouseEvent) {
+        this._lastMoveEvent = e;
+        this._pos = DOM.mousePos(this._el, e);
+
         if (!this.isActive()) {
             this._active = true;
-            this._map.moving = true;
             this._fireEvent('rotatestart', e);
             this._fireEvent('movestart', e);
             if (this._pitchWithRotate) {
@@ -143,34 +152,41 @@ class DragRotateHandler {
             }
         }
 
-        const map = this._map;
-        map.stop();
+        this._map._startAnimation(this._onDragFrame);
+    }
 
-        const p1 = this._pos,
-            p2 = DOM.mousePos(this._el, e),
+    _onDragFrame(tr: Transform) {
+        const e = this._lastMoveEvent;
+        if (!e) return;
+
+        const p1 = this._previousPos,
+            p2 = this._pos,
             bearingDiff = (p1.x - p2.x) * 0.8,
             pitchDiff = (p1.y - p2.y) * -0.5,
-            bearing = map.getBearing() - bearingDiff,
-            pitch = map.getPitch() - pitchDiff,
+            bearing = tr.bearing - bearingDiff,
+            pitch = tr.pitch - pitchDiff,
             inertia = this._inertia,
             last = inertia[inertia.length - 1];
 
         this._drainInertiaBuffer();
-        inertia.push([browser.now(), map._normalizeBearing(bearing, last[1])]);
+        inertia.push([browser.now(), this._map._normalizeBearing(bearing, last[1])]);
 
-        map.transform.bearing = bearing;
+        tr.bearing = bearing;
         if (this._pitchWithRotate) {
             this._fireEvent('pitch', e);
-            map.transform.pitch = pitch;
+            tr.pitch = pitch;
         }
 
         this._fireEvent('rotate', e);
         this._fireEvent('move', e);
 
-        this._pos = p2;
+        delete this._lastMoveEvent;
+        this._previousPos = this._pos;
     }
 
     _onUp(e: MouseEvent | FocusEvent) {
+        if (e.type === 'mouseup' && e.button !== this._eventButton) return;
+
         window.document.removeEventListener('mousemove', this._onMove, {capture: true});
         window.document.removeEventListener('mouseup', this._onUp);
         window.removeEventListener('blur', this._onUp);
@@ -180,6 +196,9 @@ class DragRotateHandler {
         if (!this.isActive()) return;
 
         this._active = false;
+        delete this._lastMoveEvent;
+        delete this._previousPos;
+
         this._fireEvent('rotateend', e);
         this._drainInertiaBuffer();
 
@@ -191,7 +210,6 @@ class DragRotateHandler {
             if (Math.abs(mapBearing) < this._bearingSnap) {
                 map.resetNorth({noMoveStart: true}, { originalEvent: e });
             } else {
-                this._map.moving = false;
                 this._fireEvent('moveend', e);
             }
             if (this._pitchWithRotate) this._fireEvent('pitchend', e);
@@ -236,8 +254,8 @@ class DragRotateHandler {
         }, { originalEvent: e });
     }
 
-    _fireEvent(type: string, e: Event) {
-        return this._map.fire(type, { originalEvent: e });
+    _fireEvent(type: string, e: ?Event) {
+        return this._map.fire(type, e ? { originalEvent: e } : {});
     }
 
     _drainInertiaBuffer() {