+++ src/style-spec/CHANGELOG.md
@@ -1,3 +1,16 @@
+
+## 11.1.1
+
+### 🐛 Bug fixes
+
+* Update SDK support information for `text-font`, and `heatmap-color`, and `hillshade` layer properties.
+
+## 11.1.0
+
+### ✨ Features and improvements
+
+* Avoid use of `new Function` for filters by converting old-style filters to expressions [#5665](https://github.com/mapbox/mapbox-gl-js/pull/5665).
+
 ## 11.0.0
 
 ### ⚠️ Breaking changes